# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API with Swagger",
        "description": "RESTFULL API sistem penjualan documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Service product",
                "description": "Endpoint get product",
                "operationId": "e68b1eace49fb57d47cb6e34def392ca",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Service product",
                "description": "Endpoint add new product",
                "operationId": "f6c6fa48346dd08aae3060e1e8765913",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "unit_id": {
                                                "type": "integer"
                                            },
                                            "category_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "desc": {
                                                "type": "string"
                                            },
                                            "price_buy": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "price_sell": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "qty": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "unit_id": 1,
                                    "category_id": 1,
                                    "name": "Beras",
                                    "desc": "Beras Pulen",
                                    "price_buy": 15000,
                                    "price_sell": 18000,
                                    "qty": 180
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Field product must be filled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Service product",
                "description": "Endpoint get product specified id",
                "operationId": "3a462c6271dfd581c8c589efd97cbde4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Service product",
                "description": "Endpoint add new product",
                "operationId": "34913f3e4a52dfddcf956f5c1411beb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product data id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "unit_id": {
                                                "type": "integer"
                                            },
                                            "category_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "desc": {
                                                "type": "string"
                                            },
                                            "price_buy": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "price_sell": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "qty": {
                                                "type": "ineteger"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "unit_id": 1,
                                    "category_id": 1,
                                    "name": "Beras",
                                    "desc": "Beras Pulen",
                                    "price_buy": 15000,
                                    "price_sell": 18000,
                                    "qty": 180
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Field product must be filled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete existing product",
                "description": "Deletes a record and returns no content",
                "operationId": "delete product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}